Re-introduce fix for domains referred to by array eltTypes
See test/users/bugzilla/bug794133/ and commit 6f435f19a800.

The code in question is code like this:

   var D = {1..0};   // start empty; we'll resize later
   var A: [D] [1..2] real;

To make that work, I simply added back in a counter in BaseDom
for the number of arrays referring to that domain in the array eltType.
Thus, the fix remains philisophically similar to the previous fix,
although it has some minor differences to accomadate the general removal
of array reference counting. This counter is currently protected
by the _arrsLock since it is often used at the same time as
_arrs.