Add logic to resolveNormalCall to support super.init() when super == object, etc
Prior to this change, I had a ton of failures which were due to object's
default constructor being _construct_object, when super.init() calls within the
body of initializers that inherited from it expected to have a init() function
to call.  Similarly, if the class being inherited from did not define any
constructor or initializer, we would encounter this issue.  This problem is a
temporary one, due to not having initializers implemented in a strong enough
state that they can be used as the compiler generated default for a class.  So
I inserted a temporary replacement that occurs when the type having init called
upon it did not define either an initializer or a constructor (and an error when
the type having init called upon it defined a constructor).

With this change, 17 new initializer futures now pass.