Streamline the named function interface
Other than the get functions, eliminate the _si and _ui qualifiers on
function names. (since chapel can overload). (other than the get
functions because the get functions need to know what they are
returning. Could have discussion on if this should be done through an
argument instead of through the name)

Update the relevant tests