Add an optimization to replace repeated array accesses with ref temps
Add a pragma "removable array access" and apply it to the array accessor (this)
functions in ChapelArray.

Within for loops, look for calls to functions that have the new pragma and
use the loop index variable as the index argument.  If more than a threshold
of these calls exist within a loop, assign one call to a reference variable and
replace the other calls with the reference.

For example:
  for i in rng {
    A[i] = A[i] * A[i];
    B[i] = A[i] + f(A[i]);
  }

Becomes:
  for i in rng {
    ref tmp = A[i];
    tmp = tmp * tmp;
    B[i] = tmp * f(tmp);
  }

This is controlled by the option --[no-]replace-array-accesses-with-ref-temps.
It is off by default and --fast and --baseline do not modify it at the moment.
Once it has been tested further, --fast will enable it and --baseline will
disable it.