Avoid Travis failure
With warnings treated as error, the previous change failed because it didn't
recognize that INT_FATAL would cause execution to finish and so was worried that
the character array "str" was not initialized before a later code block.  To
avoid that warning, just set the str variable to the empty string, since it
doesn't matter anyways.