Enable histograms using arrays with reduce intents
This change allows the user to write forall loops
with array variable(s) passed by + reduce intent:

    // A is an array
    forall .... with (+ reduce A) { .... }

Furthermore, this code will manage memory properly.

This enables a natural way to compute a histogram in parallel like this:

  var histo: [1..nBuckets] int;

    forall datum in <data to plot> with (+ reduce histo) {
    ....
    histo[the appropriate bucket] += 1;
  }

MODULE CHANGE

I modified chpl__sumType() to support the case
where the argument is an array.

Having a special case in chpl__sumType() is dissatisfying
in that ideally it would not be needed. However I did not see
worthy alternatives.  Indeed:

* (x+x).type produces an iterator record type - because here the + operator
is promoted. An iterator record type is not suitable.

* Another alternative due to Michael is:

    proc chpl__sumType(type eltType) type
    { var x: eltType;  var y = x+x; return y.type; }

  This downside of this more elegant code is that it will actually
perform the array addition at run time, incurring extra overhead.
BTW both my version and this version create the 'x' array at run time,
under the current compiler.

One issue is how this code will support skyline arrays.
Seems like if 'eltType' is the type of a skyline array,
then it should be possible to return the type of their sum as well.

Future work: we may want chpl__sumType() to report an explicit error when
two eltType-typed things cannot be added. Without that, the compiler would
generate an "unresolved call" error, which would be less clear to the user.

Note: chpl__sumType() does not have a counterpart for the other reduction
operations. This is because it allows reduction of wider variety of things
than, say, built-in multiplication.

COMPILER CHANGE

Some changes in the compiler are necessary to insert additional autoCopy
calls to prevent premature deletion, and also additional autoDestroy calls
to prevent memory leaks. Here are details.

Turns out that when I generate an AST that MOVEs the result of an
array-returning function without autoCopy-ing it, then
returnRecordsByReferenceArguments() on this function does not fire
and things go off rail. Namely, the function itself and I think other
array-returning functions that it invokes get an extra autoDestroy
on the return value. This was the case for two 'svar's,
named shadowVar and shadowVarReduc, that stored the result of identity():

// before this change:

  redRef1->insertBefore("'move'(%S, identity(%S,%S))", // init
                        svar, gMethodToken, parentOp);

Adding autoCopy on the result resolves the issue.

As well as I replaced a PRIM_ASSIGN with a call to "=" on the result
of generate(). Since the reduction var already exists, we really want
assign into it, not bitcopy. To call "=", I had to put in "new CallExpr"
explicitly. Before I relied on insertAfter() to invoke new_Expr,
alas the latter does not handle calls to functions with operator names
like "=".

NEW TESTS

The two new tests are identical, except histogram-1 has everything
at the global level and histogram-2 has all except config consts
as local code within a function (main()).