Fix problem with using an enum in a library module
Preston discovered that if you used an enum in a library module, you would get
the error "Can't find standard module 'nameOfEnum'" instead of handling the enum
use like we do in user code.  This is because standard modules are searched in a
different manner from user modules.  Since it seems reasonable to want to use an
enum in the library modules, I removed the error case from stdModNameToFilename
and returned the NULL when the name doesn't work.  I then had the parser's
parseDependentModules function continue through the loop if the filename wasn't
found, letting the potential error when neither an enum nor a standard module
was named wait until scopeResolve handles unresolved names.  This allowed his
case to work, and I verified that I still got an error when the name didn't
match.