Minor rearrangement in prep for phase 1 work
I decided that my initializers rules should live in a separate file from the
remainder of function resolution.  This would also enable me to move the call to
these checks to a different point during compilation if I so chose.  It also
will avoid blowing up functionResolution.cpp any further.

As a first step, I moved the temporaryInitializerFixup function into this new
file, added a make rule for it, and added a header file.  These will get
populated further in this branch.

Verified no change in the initializers test directory.