Removed a misleading warning about forall intents
The compiler used to warn when a record with an array field
were passed into a forall loop by default forall intent,

The warning claimed that in the following scenario:

    record MyRecord {
      var arrField: [....] int;
    }

    var outerRec: MyRecord;

    forall .... {
      .... outerRec.arrField ....;
    }

'outerRec.arrField' inside the forall referred to a *copy* of that array field
outside the forall. Whereas the current rules require that the two arrays be the same.

I verified that the two arrays are indeed the same, and removed the warning.

I added a test that confirms same-ness of the arrays, observing it under
the current default-intent rules. I also added a future illustrating incorrect
implementation of the current rules for integer fields of records.