Work-around issue with => capturing returned array
This simple test shows the problem:

  const targetLocales => setupLocaleGrid();

  proc main() {
    var copy = targetLocales;
    writeln(copy);
  }
  proc setupLocaleGrid() {
    return reshape(Locales#1, {0..#1, 0..#1}); // does not work
    //return reshape(Locales, {0..#1, 0..#1}); // neither does this
  }

This test does not currently work with the removal of array
reference counting. I'm not sure what the issue is, but since
arrays now return by value, there is no reason to use the '=>'
syntax when initializing targetLocales. Replacing it with '='
allows the test to function and so that is what I did in these
two HPL variants.