Reorganize the body of initializers
Separate the body of an initializer into three BlockStmts - phase 1, the init
call, and phase 2.

Implementation details - because phase 2 is the default, walk the FnSymbol's
body backwards from the end until we either reach the start of the body or we
find a super or this init call, whichever comes first.  Remove any code
encountered before that from the body and insert it into the phase 2 block
statement.  If an init call is encountered, move that and its supporting
statements into the init call block statement.  Then, insert the init call and
phase 2 block statements into the end of the phase 1 block statement, and
restore the return statement as the last statement in the phase 1 block.

Should not affect the current behavior yet.

Note that this traversal is too simple to handle .init() calls within loops, if
statements, and other block statements yet.  Since that is a more complicated
step, I will wait to perform it until I have the basic case working.