ForallExpr and reductions of forall expressions
* Introduce a simple ForallExpr node, used only during parsing.

* Convert certain reductions of forall expression to forall loops
with reduce intents.  Zippered foralls and filtering predicates
fall back on the old implementation.

* Due to the conversion, in the following:

    var x ...;
    + reduce [...] ... x ...

x is now passed into the corresponding forall loop by default intent.
Before this change x would refer to the outer variable directly.

ForallExpr node
---------------

This is my baby step towards my goal of representing forall loops via
dedicated AST class(es) from parser to resolution. My motivation for
the latter is to facilitate:

* a robust representation of forall intents, and

* a representation of parallel iterator invocations
that is easier to deal with.

I implemented the ForallExpr node just enough for its use during parsing.
Given that ForallExpr nodes are currently not accessed outside parsing,
additional code would be untested.

I defined ForallExpr in a simplistic manner. buildForallLoopExpr()
used to build a _parloopexpr function right away. Now it now stashes
all the pieces of information into a ForallExpr node. The _parloopexpr
function is now built in exactly the same manner - using the code in the
former buildForallLoopExpr() without modification - after all files
have been parsed, in convertForallExpressions().

So the AST for a forall expression is the same before and after this change.
With the exception of the case where a forall expression feeds into
a reduce expression. In this case, a ForallExpr node is constructed
then destructed into its pieces right away.

Conversion of reductions of forall expressions
----------------------------------------------

Before this change, a reduce expression like this:

    + reduce [i in Dom] A(i,5)

would follow the old-reduction-implementation code paths in the compiler.

Now it will be handled using the new reduction implementation
by being converted to:

    var result...;

    forall i in Dom with (+ reduce result) { result += A(i,5); }

Note in particular that the two parts of the original forall expression
are now separated into the heading and body of the resulting forall loop.

Cf. for a reduction of a non-forall expression, e.g.

    + reduce A*B

all of it ends up in the forall loop heading:

    forall elm in A*B with (+ reduce result) { result += elm; }

Test changes
------------

Added reductions over forall expressions to a couple of existing tests
because this is another way to execute forall loops.

We got a reduction-by-1 of comm counts in an SSCA2 test.