Truncate large norms and fvals for madness tests
Some of the madness tests print out values that are either extremely close to 0
or extremely large, which makes minor floating point variations across
platforms cause regressions. 3cabb63 did some work to quiet differences due to
values close to 0, but did not address differences for extremely large numbers.

To address that case, this just truncates large floating point values to the
nearest 10. I initially tried just truncating to the nearest integer, but that
wasn't enough because some of the values are almost 2 quadrillion and have
precision differences up to 0.5 (spanning an integer boundary) on different
platforms.