Make math/promotion/ tests less sensitive to floating point precision
The tests in modules/standard/math/promotion/ do some standard math operations
on arrays of reals and then print out the arrays. Unfortunately some of the
values were very close to zero, or right on integer boundaries before applying
operations, so the results varied across configurations.

For examples the test for `trunc()` was something like:

```chapel
var A = [(1.4 * 5.0 + -7.0), (1.4 * 1.0 + -7.0)] // expects [0.0, 7.0], got [~0.0, 6.999...]
trunc(A);
writeln(A); // .good expects [0.0, 7.0], but got [-4.44089e-16, 6.0] in some configs
```

Because of precision differences with either the starting array calculation, or
the math operation itself, the result could differ slightly depending on the
configuration. This just updates the tests so that we round the values iff the
value is really close to the nearest integer. This ensures that we round
results close to 0 down to 0, and that we get the starting arrays that the
tests expect.

Prior to this fix, 17 of the math/promotion tests failed on:
 - chapcs* with --fast
 - linux32 with --denormalize
 - probably a bunch of other configurations