Remove unused ?t from function signature
Later on in the function, the code explicitly checks that types of the
arguments match and produces an error if not. So, if the compiler that
all of the arguments had type t or used the :?t as reason not to choose
this function in resolution, we would get worse error messages.