Disallow certain slices of a Dimensional array
To facilitate Michael's progress towards the new array representation,
I am disabling the functionality of slicing a Dimensional array
with a domain whose type differs from that of the array's domain.
This situation is unlikely to appear in user code.

The now-disabled scenario causes problems in Michael's work
because he needs initCopy on an array to produce an array
of the same type as the argument. However, in such a scenario,
in this code:

  var A: MyTroublesomeSlice.type;

we get A._value.type != MyTroublesomeSlice._value.type,
which does not work for him.

The two types are different because each Dimensional array
keeps two domains: 'dom', which is its domain as observed
by the user, and 'allocDom', which is the domain of the array
that this slice is from. So, for example, if

  B.domain == {1..4}

then (B[1..4 by 2]) will have dom == {1..4 by 2} // stridable
and dom={1..4} // not stridable.

This makes me wonder whether Brad's upcoming array views
are taking a similar approach.

IMPLEMENTATION

The disabling is implemented by an explicit check
in Dimensional's dsiSlice().

While there, I made DimensionalArr.allocDom a non-generic field,
so that instantiated DimensionalArr types print out more tidy.

INTERESTING ASIDE

Before this change, if S were a troublesome slice, then S.type was not
troublesome any more.  I.e. the types of DimensionalArr's 'dom' and
'allocDom' differ in S._value.type while being the same in S.type.

That is because the compiler rewrites S.type into
  chpl__convertValueToRuntimeType(S)
which simply returns
  chpl__buildArrayRuntimeType(S.domain, S.eltType);
i.e. S.allocDom is not referenced anywhere.

TEST CHANGES

Two of our tests exercise this functionality: s8.chpl and s9.chpl.
I commented out the corresponding piece of these tests
and instead added a .future.

When this feature is reinstated, those pieces of s8.chpl and s9.chpl
should be un-commented and the .future be eliminated altogether.
Indeed, sx8.chpl is a subset of s8.chpl after the un-commenting.