Allow libunwind to be used in multilocale builds
When a multilocale runtime is built, the launcher is built and the
launcher builds error.c. However, the launcher does not include the
third-party libunwind headers in its build line (and it probably
shouldn't, since we'll configure libunwind for the target and the
launcher runs on the host).

This change simply adjusts error.c to require both CHPL_DO_UNWIND is set
and LAUNCHER is not set in order to activate the unwind functionality. In
order to do so, I replaced the uses of CHPL_DO_UNWIND with
CHPL_UNWIND_NOT_LAUNCHER and used #ifdefs to set
CHPL_UNWIND_NOT_LAUNCHER.