Fix --llvm --llvm-wide-opt compilation problem
One place in expr.cpp needed to be adjusted to pass an integer instead of
a char* for filename/number arg to the runtime.  Apparently this code
path does not run without --llvm-wide-opt. I adjusted the other place
that LLVM generates a file number to also generate a 32-bit value.