Implement start of Phase 1 analysis - detecting out of order initialization
This implements a check that must be performed on Phase 1 of the initializer
body - verifying that the initialization order of fields is in line with their
declaration order.

Details:
I had normalize's change_method_into_constructor pass in the type for the
initializer, since that information is not easily accessible otherwise (the
return type hasn't been determined yet, we'd have to search for the type of the
'meme' argument, and the _this field on the function takes its type from the
'meme' argument and so doesn't know its type yet).

In initializerRules.cpp, I iterate through the top level statements of the
function body, looking for obvious assignments into fields.  If I see one that
could be a field access, I verify that the field is later in the field
declarations than the previous one we saw - if it isn't a field at all, I also
will generate an error, since method calls are not allowed during Phase 1.