Fix obvious (in retrospect) error in phase 1 analysis
In thinking that field.head will always store a DefExpr, I forgot that in the
case where there were no fields it would be NULL.  Though I don't expect users
to define records or classes without fields in practice, it's still a good case
to be sure I handle correctly.  Remove the assert from the beginning of the
function, make a check on the number of fields present when we know we have a
"this" access and if that number is zero then by default the user is making a
method call too early and that isn't allowed.  If it isn't zero, *then* assert
that field.head will always store a DefExpr.