Implement detection of multiple initializations of same field in Phase 1
We decided that we prefered having a separate error message for the case where
a field is initialized multiple times in Phase 1, rather than reusing the "out
of order" error message.

Implement this by having an array of booleans indicating which fields have been
seen and if the field examined is one defined in a previous iteration,
generating a different error message.