Quote characters magic to the shell.
The GASNet amudprun launcher (currently) uses system(3) to launch the
target process(es), and thus fires initiates them through the shell.
For this reason, we need to backslash-escape any characters that are
magic to the shell in the ev var values.  (The previous solution just
silently dropped on the floor any var with a backquote in its value,
which seems cruel.)