After much turmoil and angst, fix bug with internal documentation clean up
Prior to this change, our fixInternalDocs script would remove functions and
variables with "chpl_", "_" and non-alphabet character starts, but it would fail
to remove any associated documentation comments, thus making it appear that the
symbol listed before the removed symbols had longer (and inaccurate)
documentation than was actually the case.  This change extends the removal to
also encompass the comments that followed the removed symbol.

Many thanks to Greg for the brainstorming and Ben A. for the actual awk command
structure.