Further clean ups from Mike
Add a new function, handleField, which we call from inside phase1Analysis once
we've determined that an initialization operation to something on "this" is
occurring.  This function then checks if the name being used on "this" is:

- a field we've already seen an initialization for (duplicate initialization is
not allowed)
- a field defined before the last field we saw initialization for (and thus an
out-of-order initialization, which is not allowed in phase 1),
- a field after the last field we saw (a valid operation, even if it's not the
immediate next field),
- or not even a field listed on the type (likely either a parent field, or a
method on the type of some sort).