Prevent initialization of parent fields during phase 1 of child
During phase 1, parent fields are not initialized.  Parent fields should only
be initialized within the parent initializer.  It is perfectly valid for the
child type to want a specific value in the parent field - however, if they wish
to ensure this value is there, they should set the field to this value during
Phase 2 of the initializer.

This change allows the just-added test to pass.