Make reservedSymbolsName.h depend on Makefile
This is a conservative Makefile change to avoid frustration/noise.
If the rule for making reservedSymbolsName.h changes, we need to
rebuild it.  Of course it may be the case that the Makefile will
change for other reasons, but the time to re-build this header is
not long (and other things will presumably be getting built if the
Makefile does change), and it seems more attractive to have things
be bulletproof than to require people to make clean and rebuild or
the like...