Fix block cyclic array accesses bug for 3D+ arrays
Engin sent me the test in this commit and said it doesn't seem to be working.

It turns out that the flattening logic in block cyclic wanted to walk
backwards from rank to 1, but it was written as: "rank..1 by -1".  Oops,
empty range.

With that fixed, the indexing into each block was going past the end of the
block, so tracked down another issue with mapping the N-D index to a 1-D
position in the block.