Fix ref-after-use bug in target program command line arguments.
If we take the if(c) true branch we do ADMUDP_free(dup) at the end of it
and then strcpy(end,p) right after the whole if-stmt, but p points into
dup so we're referencing already-freed memory.  Fix this by waiting
until after the strcpy() to free dup.

While here, remove the if(c) check and its else branch entirely, since
we've already done if(!c)return a few lines earlier in this function,
and  c is not modified after that.