Reorganize again.
This set up now is more optimal (because after we've seen the first field we
won't iterate more than the number of fields unless an error has occurred and
we need to verify that the name provided wasn't for an earlier field).  It does
mean that if there is an omitted, field, we will perform the iteration to verify
that multiple times (which I will likely fix, I just wanted to make sure
everything else worked first).

With this set up, inserting the initialization of omitted fields should be very
easy.