Fix indentation levels for HTML dumps when viewed with Chrome
The HTML dumps used the Definition List tag "<DL>" to increase the indentation
level, but Chrome only seems to pay attention to the first couple of these.

If the Definition Description tag "<DD>" is also used, Chrome pays better
attention to it.  So replace "<DL>" with "<DL><DD>" and replace "</DL>" with
"</DD></DL>".

This dramatically improves how these logs render in Chrome on Windows and Mac,
and I verified that it doesn't negatively impact Firefox or Safari on Mac, and
Firefox or Internet Explorer on Windows.