Mark some chpl__autoDestroy fns with the pragma
Adds pragma "auto destroy fn" to some autoDestroy functions declared in
ChapelBase. Adding this flag at the time the function is declared helps
to avoid some ordering issues with when the flag is set in function
resolution.

This change is useful on my arrays branch but should be harmless
on master.