Enable certain casts from params to array types.
We used to be unable to execute the cast 0:T
where

  type T = [....] int(8);

The reason is that the cast ends up calling chpl__transferArray(),
which assigns to elements of a T array in a forall loop.
By that time, module code forgets that we have a param
that can be converted to an int(8) - it thinks that we have an int(64),
and refuses to assign it into int(8) array elements.

To address that, I added versions of _cast() and chpl__transferArray()
that preserve param-ness of the value being assigned.

This enables _prod_id(), _bor_id(), _bxor_id() on array types like T above.

NOTES

* The _cast() I added is a counterpart to a function that's generated
automatically by build_record_cast_function() in the compiler.

* The param version of chpl__transferArray() I added contains only a single
forall loop. Everything else in the non-param chpl__transferArray()
contains zippered loops. They are not going to work with param values,
so I left those out. (Todo: what if the param value is a string?)

* I did not add a param-enabled version of the = operator (proc =)
because that would break the compiler. Indeed, the compiler expects
assignments to have 2 formals, e.g. in replaceSimpleAssignment(),
whereas a proc =(ref a, param b) becomes a 1-argument function
after resolution.
  Instead I inlined the would-be proc =() into _cast() manually.