Initial attempt at inserting omitted initialization of fields
This is going to need to change.  I try to store the original initialization
information for a field instead of the transformed version that is inserted into
the default constructor argument list, but right now at least some of what I
store gets cleaned up because it is not referenced within the AST at the end of
the pass.  I need to either figure out how to make those nodes persist or find
an alternate solution.