Remove a premature optimization which was causing incorrect behavior.
This code was causing standard and package modules which were included by
default to not allow their functions to be accessible in a qualified form.  I
believe it is because the conditions it was relying on were not maintained
(likely because this assumption was not documented where it supposedly
happened).  Or perhaps this bug has been present since the code was added.  Hard
to say.  In any case, it was originally added as a performance optimization for
visible function lookup, but since the optimization is causing incorrect
behavior, I figured it should be removed.

See commit 655529a37e5d5bc10f68b99500f4482682efd14f