Add a .future showing a memory leak
Totally generic fields in records/classes leak memory.

I believe an erroneous autoCopy is being added in buildDefaultWrapper
(around line 297, in setting the variable copyTemp). Adding
FLAG_INSERT_AUTO_DESTROY to this variable does not help because that flag
is removed in cullForDefaultConstructor.

The problem is that scopeResolve creates a default constructor that
already calls initCopy when setting the fields.

Simply removing the auto copy (on my arrays branch) led to core dumps.