Mark param c_string functions as compiler generated
This avoids a resolution issue (showing up in primers/associative.chpl
on the arrays branch).

Since a param string can coerce to a param c_string, we'd prefer
the string versions if they are usable. So, mark the param c_string
versions as "compiler generated" in order to avoid them if possible.