Add a future that involves a ref to an array element
This .future takes a reference to an array element, updates that element
through an array access, then looks at the value of the ref.  It is similar
to refToArrayElt.future in the same directory, except this one only uses a
single task.

If the ref is also written to (like in the f(true) case), then the correct
value is printed.