Enable reduce intents with arrays
This is a continuation of #4378, enabling code like this:

    var A: .... some array ....;

    forall .... with (XXX reduce A) {
      ... modify A ...
    }

for various reduction operations XXX.

Tally Of Reduce Intents On Arrays
---------------------------------

* sum: available as of #4378

* multiply: implemented in this PR

* min, max: not available
  because at present one can't invoke the min/max functions on arrays

* logical and, or: not available
  because they are currently not supported on arrays
  I think it would not be straightforward to add them on arrays

* bitwise and, or, xor: implemented in this PR

* minloc, maxloc: future work

* user-defined reduction operations: available

Memory Leaks
------------

Using a reduce intent on an array leads to a memory leak
for the following reduction operations: multiply, bitwise or and xor,
user-defined. This leak is caused by intermediary arrays not getting
freed up and is independent of reduce intents.
The leak is exhibited by the following code:

    type HistT = [1..1] int;
    const op = new ProductReduceScanOp(eltType = HistT);
    delete op;

leak specifics:
due to HistT:
  96 bytes:   DefaultRectangularDom(1,int(64),false)
per ProductReduceScanOp instance:
 136 bytes:   DefaultRectangularArr(int(64),1,int(64),false)
   8 bytes:   array elements
  24 bytes:   listNode(BaseArr)

Implementation Notes
--------------------

Each ReduceScanOp subclass in ChapelReduce.chpl needed to change from:

    var value = identity;

to

    var value = <inline the body of identity>;

Consider the construction of a class instance used for a reduce intent
when eltType is an array.  At run time, the eltType field is still there,
containing the runtime component of the array type. During class instance
initialization, the 'identity' method is invoked to initialize the 'value'
field -- assuming the class declaration has "var value = identity;".
At that point (call to 'identity') the eltType field has not been initialized
with the runtime component of the array type, so the call, which references
that runtime component, accesses the null pointer.

When the class declaration has 'identity' method inlined as in the second
version above, that does not happen. I did not investigate it in more detail.
We can come back to this issue once we ramp up with the new initializer story
implementation.

Separately, I changed the definition of _band_id() in two ways:

* Cater to array types explicitly. This makes sense because we do
not have promotion over type arguments.

* Added a compilerError to catch argument types not supported explicitly.
This required adding "else" clause to each "if", because otherwise
the compilerError call would be getting resolved - and error reported -
even for supported argument types.