Remove param c_string operations and +(c_string, c_string)
The param c_string operations will be handled by the param string operations
now that param c_string can coerce to param string and back.

The +(c_string, c_string) had a TODO next to it that it should be removed and
that module code using it should use string instead (and possibly .c_str() on
the result of the operation).

Also adds a test that c_string and string param coercions work as expected
(and that the non-param coercion is not present).