Revert "Merge pull request #4476 from lydia-duncan/defaultModule"
This reverts commit 05c25d1b7da57d3063887780c075ddb6faf37fbe, reversing
changes made to 2ca76c10b604dddaff614388f1f5a665228b7f61.

Though this implementation is buggy, we aren't willing to accept the 10% slow
down in compilation that it yields.  Ideally, this optimization will be fixed
so that it will be more closely tied to the operation it is trying to mimic,
but I don't think there is time to do that before the release.  Additionally,
we think that Ben's work to limit the scope of certain uses will lower the
likelihood of encountering this bug to the point where it is livable for the
time being (although the List module will still encounter this problem).