Insert statements for omitted field initialization
Not quite done yet.  Haven't removed the default constructor call, or ensured
that I do the right thing for records.  But still!

Stores the type for the field as well (but in a form that survives scope
resolution), then if there's a default value provided, use that, otherwise, use
a PRIM_SET_MEMBER for simple types, or a PRIM_INIT if the type is somewhat
more complicated.

Not thoroughly tested.  But the basic case worked!