Change how we touch qthreads autoconf files to match what we do for hwloc
While trying to upgrade qthreads versions I ran into an issue where the
configure step tried to rerun automake/autoconf steps. We try to prevent this
by touching certain configure/.m4/.in files in a certain order, but the way we
were doing it before wasn't quite right.

It turns out that this has been an issue for a while, which is why we sometimes
saw that qthreads files had been modified after a make. Beyond that we didn't
really notice the issue, because most of our systems happened to be using the
same version of automake that qthreads had been built with, but that's not the
case with the new qthreads version. This just updates the logic to match what
we do for hwloc, which seems to work.

This is similar to #69 but for qthreads instead of hwloc