Upgrade from qthreads 1.10 to a 1.11 release candidate
This includes the new distrib scheduler and FEB operations. This commit just
upgrades qthreads to test things out, it doesn't enable either of those new
features yet.

Note that this is a release candidate so we can start testing the new scheduler
as soon as possible, but we should have an official release in the next few
days.