Nicer signatures for compilerAssert et al.
Switch compilerError(), compilerWarning(), compilerAssert()
to accept these arguments:

* param msg:string      (varargs; optional for compilerAssert)
* param errorDepth:int  (optional)
* param test: bool      (compilerAssert only)

This allowed / forced me to get rid of a bunch of c_strings,
replacing them with strings - in the tests and the compiler.

DETAILS on compilerError, compilerWarning, compilerAssert

* Renamed the formal: "x" --> "msg", for clarity.

* Change the formal's type c_string --> string,
because we want to move away from c_strings.

* Replaced the fixed number of formals in compilerAssert() functions
(we had them up to 5 message formals) with a single vararg "msg" formal.
So now it is the same as for compilerError/compilerWarning.

Explanation: given that compilerAssert invokes compilerError,
compilerAssert's formals must preserve their param-ness
when being passed to compilerError.

This was enabled by #3930. Now each individual formal of the msg vararg
is passed directly to compilerError, so param-ness is preserved.
Prior to #3930, these individual formals would get bundled into a tuple,
so their param-ness would get lost because we do not have param tuples.

* While there, removed the actuals to "error" and "warning" primitive calls.
Those primitives do not actually consult their arguments -
they examine the formals of the enclosing function instead -
so passing the arguments is misleading.

REPLACING c_strings WITH strings

Apparently a param c_string actual is not coercible to a param string formal.
Therefore I replaced c_string with string everywhere a c_string was passed to
compilerError, compilerWarning, compilerAssert.

This included:

* in tests: compilerWarning(..., xxxxx:c_string) and
  c_string-typed param formals

* in compiler: new_CStringSymbol()
  - only when this symbol was passed to one of the compiler* functions

CHANGES TO test/users/vass/compilerAssert/compilerAssert*

* compilerAssert7* now produce the desired output.
I updated .good and removed a now-obsolete "pardon the mess" comment
from .chpl

* compilerAssert6* do not exercise any code paths through the compiler
that compilerAssert5* exercise (so I think), so I removed them.

We could say the same about compilerAssert7 and remove it - or say the same
about compilerAssert5 and remove those, keeping compilerAssert7.
I did not go that far, being extra-cautious.